sys-power/reform2-lpc-module: Improved support for dist-kernel.
The lpc module would always build for the currently running kernel.
Instead, we now build for the latest installed kernel, similar to
how other Gentoo packages do it. This enables us to rebuild the module
just after upgrading the kernel, without having to reboot first. This
in turn then allows the use of the dist-kernel USE which does this
automatically on every kernel upgrade.